+++ Assets/BulkTools/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changes:
 
+
+
+## [0.0.3]
+
+### Added:
+- Added Util Pack Module
+    - List Extensions (Shuffling)
+    - Random Extras (Sequences, Weighted Random)
+    - Coroutines With Results
+- Added Object Pool Module
+
+### Updated:
+- Updated Experimental tagging on modules that have not undergone sufficient testing
+
+
+
 ## [0.0.2]
 
 ### Fixes:
@@ -11,6 +27,7 @@
 ### Removed:
 - Removed Better Tags dependency on TestModule
 - Removed TestModule
+
 
 
 ## [0.0.1]